requirements.txt
asgiref==3.11.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
diff-match-patch==20241021
Django==6.0.2
django-allauth==65.14.3
django-filter==25.2
django-import-export==4.4.0
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.11.0
requests==2.32.5
sqlparse==0.5.5
stripe==14.3.0
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
